+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### `@stlite/desktop`
+
+#### Added
+
+- A custom `package.json` field `stlite.embed` has been added to control the embed mode, [#585](https://github.com/whitphx/stlite/pull/585).
+
 ## [0.35.1] - 2023-07-03
 
 ### `@stlite/desktop`